challenges/AI/Boxes/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
datasets==2.19.0
dill==0.3.8
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2024.3.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==65.5.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
fastapi==0.103.2
pydantic==2.6.2
uvicorn[standard]==0.23.2
